+++ scene.h
@@ -28,4 +28,4 @@ scene_t *scene_init(uint32_t w, uint32_t h, rgba_t *canv, uint8_t *zbuf);
 void scene_free(scene_t *s);
 zbuf_t *zbuf_init(int w, int h, uint8_t *data);void clear(scene_t *s);
 void frame(scene_t *s);
-int find_line_x(point_t a, point_t b, int y);
+int16_t find_line_x(point_t a, point_t b, int16_t y);
